// golden-route-react/src/interfaces/flightInterface.tsx
export interface Flight {
    icao24: string;
    callsign: string;
    origin_country: string;
    time_position: number;
    last_contact: number;
    longitude: number;
    latitude: number;
    baro_altitude: number;
    on_ground: boolean;
    velocity: number;
    true_track: number;
    vertical_rate: number;
    sensors: number[];
    geo_altitude: number;
    squawk: string;
    spi: boolean;
    position_source: number;
    category: number;
    closingTime: number | null; 
    src?: string; 
}
  
export  interface FlightBrief {
    latitude: number;
    longitude: number;
    callsign: string;
}